\name{getCounts}
\alias{getCounts}
\title{Given a subsetted dataframe from the Global Knowledge Graph, return the info in the "Counts" field as a data frame.}
\usage{
  getCounts(gkg.df)
}
\arguments{
  \item{gkg.df}{\code{gkg.df}}
}
\value{
  counts A data frame containing count information.
}
\description{
  This will only give you the info in the counts field, and
  in no particular order. Next steps: getting the date and
  themes to come with it.
}
\examples{
R code here showing how your function works
}
\keyword{GDELT,}
\keyword{gdeltr}

