\name{nameFixer}
\alias{nameFixer}
\title{Standardize names in a data frame.}
\usage{
  nameFixer(namevector)
}
\arguments{
  \item{data}{A vector of names \code{gkg.df}}
}
\value{
  newvec A vector with cleaned names
}
\description{
  This will be very much a function in progress.  The focus
  is on Syrian names right now. Credit to
  http://susanejohnston.wordpress.com/ for find-and-replace
  code.
}
\examples{
Some R code here
}
\keyword{GDELT,}
\keyword{gdeltr}

