\docType{data}
\name{nameFixer_data}
\alias{nameFixer_data}
\title{Name translation dataframe for GKG}
\format{A data frame with a few dozen rows and 2 columns}
\description{
  \itemize{ \item oldvalue: The name to be replaced \item
  newvalue: The standard name form }
}
\note{
  This is a work in progress.
}
\keyword{datasets}
\keyword{datasets,}
\keyword{gdelt,}
\keyword{gdeltr}

